Update pkidestroy and pki-server remove
pkidestroy and pki-server remove commands have been modified to
work more consistently. Now by default the logs will be retained
so the following folders will remain:

- /var/lib/pki/<instance>
- /var/lib/pki/<instance>/logs
- /var/log/pki/<instance> for instances using FHS layout

The tools will also provide an option to remove the logs if they
are no longer needed.

The tests have been updated accordingly.

+++ base/server/python/pki/server/__init__.py
@@ -1142,7 +1142,7 @@ grant codeBase "file:%s" {
 
         logger.info('Web application stopped')
 
-    def remove(self, force=False):
+    def remove(self, remove_logs=False, force=False):
 
         logger.info('Removing %s', self.service_conf)
         pki.util.remove(self.service_conf, force=force)
@@ -1156,17 +1156,22 @@ grant codeBase "file:%s" {
         logger.info('Removing %s', self.temp_dir)
         pki.util.rmtree(self.temp_dir, force=force)
 
-        logger.info('Removing %s', self.log_dir)
-        pki.util.rmtree(self.log_dir, force=force)
+        if remove_logs:
+            logger.info('Removing %s', self.log_dir)
+            pki.util.rmtree(self.log_dir, force=force)
 
         self.remove_libs(force=force)
+
         self.remove_conf_dir(force=force)
 
         logger.info('Removing %s', self.bin_dir)
         pki.util.unlink(self.bin_dir, force=force)
 
-        logger.info('Removing %s', self.base_dir)
-        pki.util.rmtree(self.base_dir, force=force)
+        if os.path.isdir(self.base_dir) and not os.listdir(self.base_dir):
+
+            # Remove instance base dir if empty
+            logger.info('Removing %s', self.base_dir)
+            pki.util.rmtree(self.base_dir, force=force)
 
     def remove_libs(self, force=False):
